rag-visualizer/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
tiktoken
pydantic
websockets
